+++ docs/changes.md
@@ -1,6 +1,13 @@
 # dev
 Release date: `2017-??-??`
 
+### Core
+- [NXDRIVE-903](https://jira.nuxeo.com/browse/NXDRIVE-903): Renaming folders/files does not sync while network interface is OFF
+- [NXDRIVE-904](https://jira.nuxeo.com/browse/NXDRIVE-904): Renaming folders and removing files does not sync while not running
+
+#### Minor changes
+- Utils: `guess_server_url()` now checks for the good HTTP status code
+
 
 # 2.5.3
 Release date: `2017-09-06`