+++ docs/changes.md
@@ -6,6 +6,7 @@ Release date: `2017-??-??`
 
 ### Packaging / Build
 - [NXDRIVE-983](https://jira.nuxeo.com/browse/NXDRIVE-983): Windows pip installation failure because of inexistant DLL
+- [NXDRIVE-977](https://jira.nuxeo.com/browse/NXDRIVE-977): Drive-packages should fail on error
 
 #### Minor changes
 - Utils: `guess_server_url()` now checks for the good HTTP status code